Include authors' names w/featured data
For the time being, this will be only implemented on the home page featured data tile. Including this on the grid-page tiles results in loss of descriptions (i.e. not enough space to include both). A solution to this may be larger tiles, perhaps one per row, but will be implemented later unless this item receives higher priority.