Update search bar aesthetics
Applied a design common across websites. Replaced "Search Website on Text" with "Search" because searching the "website" is implied. Replaced the text of the search button with an icon because place holder text "Search" followed by a button titled "Search" is redundant.

Matched search button color to website color scheme.