Update home page links to grid pages
It was suggested that hyperlinking the section titles "Featured Data" and "News" is a common pattern we could use and which allows removal of the somewhat redundant "All Featured Data" and "All News" buttons at the bottom of each section.

This hyperlinking is accompanied by an underline hover effect on section titles to match the liveliness of the home page.