Update basic page CSS for consistency
Basic pages (without a table of contents) used CSS that resulted in a page format differing from non-basic pages. Most notably left-justified content, where other pages were centered, and inconsistent view port scaling between "wide" and "tall" screen layouts. Additionally, the content view port of the basic pages were ~ 25% larger, which in some ways is less of an issue because of the balance a table of contents provided.

To address this issue the basic.css has been refactored and applied to related basic pages. Additionally, the main content has been wrapped in a container element. Note, the EDI Products page is an edge case requiring the absence of a wrapping container element. To handle this edge case, a separate set of logic has been added to md2html() of build.py.

Fixes #27