Fix variable placement of search field
Adding style="margin-block-end: 0em" overrides the user agent style sheet, which for some reason adds a 1em to the header and footer <form> blocks. Another oddity is if <!DOCTYPE html> is added to non index.html pages then this hack doesn't need to be applied, however, doing so breaks the sticky footer. So ... we are hacking it for the time being.

Fixes #31