Add CTS logo to footer
The footer is where we link to affiliates and their websites.
Recommendations, from CTS, on what the logo in the footer of our website
should link to are ambiguous. They do recommend "placing the logo on your
repository website with the link to the public version of your
application.", which we will do from our Policy page.

To be consistent with our current pattern of linking affiliate logos in
our footer to the respective websites, and not going against the CTS
recommendations, this commit adds the CTS logo accordingly.