Add CSS for displaying certifications
The new certifications section of the projects "About" page borrows HTML
formatting used in the projects "Team" section, because we like the
right justified display of team member portraits and think this would
work well for the certification logos. However, the bottom padding of
the CSS person_row class creates an unusually large blank space below
the section, which is not present in most sections on this page.

This commit creates and applies a new certification_row class to address
the large blank space issue.