Link "Find Data" button to advanced search
Link the "Find Data" button on the home page, and drop down menu
item directly to the Advanced Search interface of the Data Portal,
rather than the intermediary page containing links to the 3 search
engines through which EDI data can be found. Do this to provide
the user a more streamlined and intuitive experience.

The original rationale for linking to an intermediary page was to
highlight the extra effort we put forth to make EDI data findable.
However, such information is provided elsewhere in the website.